tests/wpt/web-platform-tests/web-animations/animation-effect-timing/getAnimations.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Element.getAnimations tests</title>
<link rel="help" href="http://w3c.github.io/web-animations/#animationeffecttiming">
<link rel="author" title="Ryo Motozawa" href="mailto:motozawa@mozilla-japan.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
'use strict';

test(function(t) {
  var div = createDiv(t);
  var anim = div.animate({ opacity: [ 0, 1 ] }, 2000);
  anim.finish();
  assert_equals(div.getAnimations().length, 0, 'animation finished');
  anim.effect.timing.duration += 100000;
  assert_equals(div.getAnimations()[0], anim, 'set duration 102000');
  anim.effect.timing.duration = 0;
  assert_equals(div.getAnimations().length, 0, 'set duration 0');
  anim.effect.timing.duration = 'auto';
  assert_equals(div.getAnimations().length, 0, 'set duration \'auto\'');
}, 'when duration is changed');

</script>
</body>
